# openapi.yaml
openapi: 3.0.3
info:
  title: AI Agent API
  description: |
    API pour le traitement automatisé des documents d'identité, permis de conduire et cartes grises avec intelligence artificielle.
    
    ## Fonctionnalités
    - Traitement OCR automatisé avec Azure Cognitive Services
    - Extraction intelligente des données structurées
    - Support multilingue (Français/Arabe)
    - Analytics et rapports en temps réel
    - Authentification JWT sécurisée
    
    ## Workflow Typique
    1. **Inscription/Connexion** : Créer un compte et récupérer un token JWT
    2. **Upload Documents** : Envoyer les images des documents (recto/verso)
    3. **Traitement IA** : L'API extrait et structure automatiquement les données
    4. **Récupération** : Accéder aux données structurées et aux analytics
    
    ## Support
    - Documentation interactive disponible sur `/docs/`
    - Exemples complets dans le README.md
    - Support technique : support@ai-agent.com
  version: 1.0.0
  contact:
    name: Support API AI Agent
    email: support@ai-agent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Serveur de développement
  - url: https://api.ai-agent.com
    description: Serveur de production
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentification
      summary: Inscription d'un nouvel utilisateur
      description: Créer un nouveau compte utilisateur dans le système
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: Adresse email de l'utilisateur
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  description: Mot de passe (minimum 6 caractères)
                  example: motdepasse123
                full_name:
                  type: string
                  description: Nom complet de l'utilisateur
                  example: Jean Dupont
      responses:
        '201':
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentification
      summary: Connexion utilisateur
      description: Authentifier un utilisateur et récupérer un token JWT
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: motdepasse123
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT Token d'authentification
                    example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /me:
    get:
      tags:
        - Profil Utilisateur
      summary: Récupération du profil utilisateur
      description: Obtenir les informations du profil de l'utilisateur connecté
      responses:
        '200':
          description: Profil utilisateur récupéré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Profil Utilisateur
      summary: Mise à jour du profil utilisateur
      description: Modifier les informations du profil utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: Nouveau Nom
                phone:
                  type: string
                  example: +212600000000
                address:
                  type: string
                  example: 123 Rue Principale
                job_title:
                  type: string
                  example: Développeur
                department:
                  type: string
                  example: IT
      responses:
        '200':
          description: Profil mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profil mis à jour avec succès
                  user:
                    $ref: '#/components/schemas/UserBasic'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cin/process:
    post:
      tags:
        - Carte d'Identité Nationale
      summary: Traitement d'une CIN
      description: |
        Traite une carte d'identité nationale marocaine en uploadant les images recto et verso.
        L'API utilise l'OCR Azure pour extraire le texte et l'IA pour structurer les données.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - recto
                - verso
              properties:
                recto:
                  type: string
                  format: binary
                  description: Image du recto de la carte (JPG, PNG, max 10MB)
                verso:
                  type: string
                  format: binary
                  description: Image du verso de la carte (JPG, PNG, max 10MB)
      responses:
        '200':
          description: CIN traitée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CINData'
        '400':
          description: Fichiers manquants ou invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur de traitement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cin/all:
    get:
      tags:
        - Carte d'Identité Nationale
      summary: Récupération de toutes les CIN
      description: Obtenir la liste de toutes les cartes d'identité traitées
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des CIN récupérée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CINData'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erreur serveur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /permis/process:
    post:
      tags:
        - Permis de Conduire
      summary: Traitement d'un permis de conduire
      description: |
        Traite un permis de conduire marocain en uploadant les images.
        Le recto est obligatoire, le verso est optionnel selon le type de permis.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - recto
              properties:
                recto:
                  type: string
                  format: binary
                  description: Image du recto du permis (obligatoire)
                verso:
                  type: string
                  format: binary
                  description: Image du verso du permis (optionnel)
      responses:
        '200':
          description: Permis traité avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermisData'
        '400':
          description: Fichiers manquants ou invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /permis/all:
    get:
      tags:
        - Permis de Conduire
      summary: Récupération de tous les permis
      description: Obtenir la liste de tous les permis de conduire traités
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des permis récupérée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermisData'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /gris/process:
    post:
      tags:
        - Carte Grise
      summary: Traitement d'une carte grise
      description: |
        Traite une carte grise (certificat d'immatriculation) marocaine.
        Nécessite les images recto et verso du document.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - recto
                - verso
              properties:
                recto:
                  type: string
                  format: binary
                  description: Image du recto de la carte grise
                verso:
                  type: string
                  format: binary
                  description: Image du verso de la carte grise
      responses:
        '200':
          description: Carte grise traitée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartGrisData'
        '400':
          description: Fichiers manquants ou invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /gris/all:
    get:
      tags:
        - Carte Grise
      summary: Récupération de toutes les cartes grises
      description: Obtenir la liste de toutes les cartes grises traitées
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des cartes grises récupérée
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartGrisData'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /gris/evolution-mensuel:
    get:
      tags:
        - Carte Grise
      summary: Évolution mensuelle des immatriculations
      description: Statistiques d'évolution mensuelle des premières immatriculations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Données d'évolution mensuelle
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                example:
                  "2024-01": 15
                  "2024-02": 23
                  "2024-03": 18
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /charts/overview:
    get:
      tags:
        - Analytics & Graphiques
      summary: Vue d'ensemble des cartes
      description: Statistiques générales sur tous les types de cartes traitées
      responses:
        '200':
          description: Vue d'ensemble récupérée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOverview'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /charts/gender-distribution:
    get:
      tags:
        - Analytics & Graphiques
      summary: Distribution par genre
      description: Répartition des utilisateurs par genre (M/F)
      responses:
        '200':
          description: Distribution par genre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenderDistribution'
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /charts/license-categories:
    get:
      tags:
        - Analytics & Graphiques
      summary: Distribution des catégories de permis
      description: Répartition des permis par catégorie (A, B, C, D, etc.)
      responses:
        '200':
          description: Distribution des catégories de permis
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
                example:
                  A: 5
                  B: 120
                  C: 15
                  D: 8
        '401':
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Token obtenu via l'endpoint `/auth/login`.
        Format: `Bearer <your_jwt_token>`
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Message d'erreur détaillé
          example: Email and password required
...
    UserBasic:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        full_name:
          type: string
          example: Jean Dupont
        id:
          type: integer
          example: 123
...